+++ client/src/pages/tools/json-yaml-converter.tsx
@@ -1,4 +1,4 @@
-import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { TextArea } from "@/components/ui/textarea";
 import { useTheme } from "@/providers/theme-provider";
@@ -32,7 +32,7 @@ import {
 const DEFAULT_YAML = convertJSONToYAML(DEFAULT_JSON).converted;
 
 export default function JSONYAMLConverter() {
-  const { toast } = useToast();
+  const { handleShare } = useShareTool();
   const tool = getToolByPath("/tools/json-yaml-converter");
   const [jsonText, setJsonText] = useState(DEFAULT_JSON);
   const [yamlText, setYamlText] = useState("");
@@ -92,22 +92,6 @@ export default function JSONYAMLConverter() {
     (yamlText !== DEFAULT_YAML && yamlText.trim() !== "");
   const isAtDefault =
     jsonText === DEFAULT_JSON && (yamlText === "" || yamlText === DEFAULT_YAML);
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
-  };
 
   return (
     <div className="max-w-6xl mx-auto">
