+++ client/src/pages/tools/jwt-decoder.tsx
@@ -1,4 +1,4 @@
-import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { TextArea } from "@/components/ui/textarea";
 import { Alert, AlertDescription } from "@/components/ui/alert";
@@ -28,7 +28,7 @@ import {
 } from "@/components/ui/tooltip";
 
 export default function JWTDecoder() {
-  const { toast } = useToast();
+  const { handleShare } = useShareTool();
   const tool = getToolByPath("/tools/jwt-decoder");
   const [token, setToken] = useState(DEFAULT_JWT);
   const [header, setHeader] = useState("");
@@ -118,22 +118,6 @@ export default function JWTDecoder() {
       decodeToken();
     }
   }, [autoProcess, decodeToken]);
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
-  };
 
   return (
     <div className="max-w-6xl mx-auto">
