+++ client/src/pages/tools/microphone-test.tsx
@@ -11,6 +11,7 @@ import {
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Mic, Play, Square, Download, Pause } from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 
 import { SecurityBanner } from "@/components/ui/security-banner";
 import { getToolByPath } from "@/data/tools";
@@ -44,6 +45,7 @@ export default function MicrophoneTest() {
   const animationFrameRef = useRef<number | null>(null);
   const isRecordingRef = useRef<boolean>(false);
   const { toast } = useToast();
+  const { handleShare } = useShareTool();
 
   const requestPermission = async () => {
     try {
@@ -388,22 +390,6 @@ export default function MicrophoneTest() {
   };
 
   const permissionStatus = getPermissionStatus();
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
-  };
 
   return (
     <div className="max-w-6xl mx-auto">
