+++ client/src/pages/tools/password-generator.tsx
@@ -17,7 +17,7 @@ import {
 
 import { useToolDefault } from "@/hooks/use-tool-default";
 import { usePersistentForm } from "@/hooks/use-persistent-state";
-import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { getToolByPath } from "@/data/tools";
 import { ToolExplanations } from "@/components/tool-explanations";
 import { ShortcutBadge } from "@/components/ui/shortcut-badge";
@@ -31,7 +31,7 @@ interface PasswordStrength {
 }
 
 export default function PasswordGenerator() {
-  const { toast } = useToast();
+  const { handleShare } = useShareTool();
   const tool = getToolByPath("/tools/password-generator");
   const { fields, updateField, resetFields } = usePersistentForm(
     "password-generator",
@@ -274,22 +274,6 @@ export default function PasswordGenerator() {
       newVisible.add(index);
     }
     setVisiblePasswords(newVisible);
-  };
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
   };
 
   return (
