+++ client/src/pages/tools/search-replace.tsx
@@ -1,4 +1,4 @@
-import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
@@ -34,7 +34,7 @@ import {
 } from "@/components/ui/tooltip";
 
 export default function SearchReplace() {
-  const { toast } = useToast();
+  const { handleShare } = useShareTool();
   const tool = getToolByPath("/tools/search-replace");
   const [text, setText] = useState(DEFAULT_SEARCH_REPLACE_TEXT);
   const [searchText, setSearchText] = useState(DEFAULT_SEARCH_REPLACE_SEARCH);
@@ -142,22 +142,6 @@ export default function SearchReplace() {
       performSearchReplace();
     }
   }, [autoProcess, performSearchReplace]);
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
-  };
 
   return (
     <div className="max-w-6xl mx-auto">
