+++ client/src/pages/tools/tls-decoder.tsx
@@ -1,4 +1,4 @@
-import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { TextArea } from "@/components/ui/textarea";
 import { useTheme } from "@/providers/theme-provider";
@@ -45,7 +45,7 @@ interface CertificateInfo {
 }
 
 export default function TLSDecoder() {
-  const { toast } = useToast();
+  const { handleShare } = useShareTool();
   const tool = getToolByPath("/tools/tls-decoder");
   const [certificate, setCertificate] = useState(DEFAULT_TLS_DECODER);
   // Incrementing key to force CodeMirror remount on reset/clear preventing residual merged content
@@ -172,22 +172,6 @@ export default function TLSDecoder() {
   };
 
   const validityStatus = getValidityStatus();
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
-  };
 
   return (
     <div className="max-w-6xl mx-auto">
