+++ client/src/pages/tools/world-clock.tsx
@@ -23,6 +23,7 @@ import {
   DataButtonGroup,
 } from "@/components/ui/tool-button";
 import { useToast } from "@/hooks/use-toast";
+import { useShareTool } from "@/hooks/use-share-tool";
 import { getToolByPath } from "@/data/tools";
 import { ToolExplanations } from "@/components/tool-explanations";
 import { ShortcutBadge } from "@/components/ui/shortcut-badge";
@@ -44,6 +45,7 @@ export default function WorldClock() {
   const [showAddClock, setShowAddClock] = useState(false);
 
   const { toast } = useToast();
+  const { handleShare } = useShareTool();
 
   useEffect(() => {
     const updateTimes = () => {
@@ -162,22 +164,6 @@ export default function WorldClock() {
       title: "Custom Clocks Cleared",
       description: "All custom clocks have been removed.",
     });
-  };
-
-  const handleShare = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      toast({
-        title: "Link copied!",
-        description: "Tool URL copied to clipboard",
-      });
-    } catch {
-      toast({
-        title: "Share failed",
-        description: "Could not copy URL to clipboard",
-        variant: "destructive",
-      });
-    }
   };
 
   return (
